Refactor PyVEX to pass lint
