+++ .pre-commit-config.yaml
@@ -78,7 +78,7 @@ repos:
     rev: v1.10.0
     hooks:
     # Python
-    # -   id: python-use-type-annotations
+    -   id: python-use-type-annotations
     -   id: python-no-log-warn
     # Documentation
     -   id: rst-backticks
@@ -92,7 +92,7 @@ repos:
     -   id: check-builtin-literals
     -   id: check-docstring-first
 
-# -   repo: https://github.com/charliermarsh/ruff-pre-commit
-#     rev: v0.0.230
-#     hooks:
-#     - id: ruff
+-   repo: https://github.com/charliermarsh/ruff-pre-commit
+    rev: v0.0.231
+    hooks:
+    - id: ruff
