+++ pyproject.toml
@@ -10,3 +10,12 @@ force-exclude = '''
   vex
 )/
 '''
+
+[tool.ruff]
+line-length = 120
+select = [
+  "E",
+  "F",
+  "I",
+  "TID",
+]
