+++ pyvex/lifting/gym/__init__.py
@@ -1,5 +1,5 @@
-from .arm_spotter import ARMSpotter
 from .aarch64_spotter import AARCH64Spotter
-from .x86_spotter import X86Spotter, AMD64Spotter
+from .arm_spotter import ARMSpotter
+from .x86_spotter import AMD64Spotter, X86Spotter
 
 __all__ = ("ARMSpotter", "AARCH64Spotter", "X86Spotter", "AMD64Spotter")
