+++ pyvex/lifting/gym/aarch64_spotter.py
@@ -1,9 +1,10 @@
 import logging
-from ..util.lifter_helper import GymratLifter
-from ..util.instr_helper import Instruction
-from .. import register
 
-l = logging.getLogger(__name__)
+from pyvex.lifting.lift import register
+from pyvex.lifting.util.instr_helper import Instruction
+from pyvex.lifting.util.lifter_helper import GymratLifter
+
+log = logging.getLogger(__name__)
 
 
 class Aarch64Instruction(Instruction):  # pylint: disable=abstract-method
@@ -17,7 +18,7 @@ class Instruction_SYSL(Aarch64Instruction):
     bin_format = "1101010100101qqqnnnnmmmmppprrrrr"
 
     def compute_result(self):  # pylint: disable=arguments-differ
-        l.debug("Ignoring SYSL instruction at %#x.", self.addr)
+        log.debug("Ignoring SYSL instruction at %#x.", self.addr)
 
 
 class Instruction_MSR(Aarch64Instruction):
@@ -25,7 +26,7 @@ class Instruction_MSR(Aarch64Instruction):
     bin_format = "11010101000ioqqqnnnnmmmmppprrrrr"
 
     def compute_result(self):  # pylint: disable=arguments-differ
-        l.debug("Ignoring MSR instruction at %#x.", self.addr)
+        log.debug("Ignoring MSR instruction at %#x.", self.addr)
 
 
 class Instruction_MRS(Aarch64Instruction):
@@ -33,7 +34,7 @@ class Instruction_MRS(Aarch64Instruction):
     bin_format = "110101010011opppnnnnmmmmppprrrrr"
 
     def compute_result(self):  # pylint: disable=arguments-differ
-        l.debug("Ignoring MRS instruction at %#x.", self.addr)
+        log.debug("Ignoring MRS instruction at %#x.", self.addr)
 
 
 class AARCH64Spotter(GymratLifter):
