+++ pyvex/lifting/util/__init__.py
@@ -1,4 +1,14 @@
-from .vex_helper import Type, JumpKind
-from .syntax_wrapper import VexValue
-from .instr_helper import ParseError, Instruction
+from .instr_helper import Instruction
 from .lifter_helper import GymratLifter, ParseError
+from .syntax_wrapper import VexValue
+from .vex_helper import JumpKind, Type
+
+__all__ = [
+    "Type",
+    "JumpKind",
+    "VexValue",
+    "ParseError",
+    "Instruction",
+    "GymratLifter",
+    "ParseError",
+]
