+++ pyvex/lifting/util/instr_helper.py
@@ -1,14 +1,13 @@
 import abc
 import string
+
 import bitstring
-import logging
+
+from pyvex.expr import IRExpr, RdTmp
 
 from .lifter_helper import ParseError
 from .syntax_wrapper import VexValue
-from ...expr import IRExpr, RdTmp
-from .vex_helper import JumpKind, vex_int_class, IRSBCustomizer
-
-l = logging.getLogger("instr")
+from .vex_helper import IRSBCustomizer, JumpKind, vex_int_class
 
 
 class Instruction(metaclass=abc.ABCMeta):
