+++ pyvex/lifting/zerodivision.py
@@ -1,7 +1,8 @@
 import copy
 
-from . import Postprocessor, register
-from . import libvex
+from pyvex import const, expr, stmt
+
+from .post_processor import Postprocessor
 
 
 class ZeroDivisionPostProcessor(Postprocessor):
@@ -69,10 +70,3 @@ class ZeroDivisionPostProcessor(Postprocessor):
 
         for i, s in reversed(insertions):
             self.irsb.statements.insert(i, s)
-
-
-# for arch_name in libvex.SUPPORTED:
-#    register(ZeroDivisionPostProcessor, arch_name)
-
-
-from .. import stmt, expr, const
