+++ tests/test_gym.py
@@ -2,6 +2,7 @@
 import unittest
 
 import archinfo
+
 import pyvex
