+++ tests/test_mips32_postprocess.py
@@ -1,4 +1,5 @@
 import archinfo
+
 import pyvex
