+++ tests/test_s390x_exrl.py
@@ -1,4 +1,5 @@
 import archinfo
+
 import pyvex
