+++ tests/test_s390x_vl.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
 import archinfo
+
 import pyvex
