+++ tests/test_ud2.py
@@ -1,4 +1,5 @@
 import archinfo
+
 import pyvex
